Add git-btrfs
This provides a nifty `git btrfs worktree` command that combines git
worktrees and Btrfs snapshots. It's useful for, say, Linux kernel
development, where you can have a worktree that you use to track and
build the latest -rc, then snapshot a branch derived from that -rc,
saving a recompilation of the whole kernel for some quick changes. It
requires that the repository be its own Btrfs subvolume, so this also
adds a `git btrfs clone` command that clones a repo into a subvolume.