git-btrfs: initialize index file with reflink copy
The `git reset` we do to rebuild the index takes several seconds because
it has to write out the entire index. However, we can initialize it with
the state of the index in the worktree we are snapshotting, which will
only differ from the final result if there is anything staged.