Add tool for fixing up objtool dependencies after git btrfs worktree
objtool in the kernel repo uses absolute paths in its dependencies,
which forces a rebuild of objtool when the path changes (e.g., because
it was renamed or snapshotted), which in turn forces a rebuild of the
entire kernel, defeating the purpose of git btrfs worktree. As a hack,
we can rewrite the dependency files to refer to the new path.