vim: improve line continuation macro
When editing in the middle of a macro, we should still add a
continuation on the last line of a selection. The same goes when only
selecting one line that already has a line continuation. Also add a
normal mode mapping which does the same thing as selecting only the
current line.