Switch to GNOME
I've been using dwm for a long time, but I've been experimenting with
GNOME for its better HiDPI support and more polished feel in general. A
non-tiling window manager will take some getting used to, and it's a
bummer to abandon inputconfd, supavolumed, and verbar, but this is
definitely less janky and hopefully less work to maintain.