Remove pythonrc
Commit 50f3484936bc ("Update pythonrc to only apply to Python 2.")
disabled pythonrc for Python3 because Python 3 configures readline by
default. Commit 773bee7a818e ("Re-enable pythonrc for Python 3.") undid
that because the default completer behaved strangely at the beginning of
a line. That was fixed in CPython commit aaf6114b37df ("Issue #23441:
rcompleter now prints a tab character instead of displaying possible
completions for an empty word.") (in v3.4.4). Python 2 is dead, as are
versions of Python 3 without that fix. So, we can finally drop this.